+++ tests/common.h
@@ -131,7 +131,7 @@ int8_t saturate_int8(int a);
 uint8_t saturate_uint8(unsigned int a);
 int16_t saturate_int16(int a);
 uint16_t saturate_uint16(unsigned int a);
-int32_t saturate_int32(int a);
+int32_t saturate_int32(int64_t a);
 uint32_t saturate_uint32(uint64_t a);
 
 float ranf(float low, float high);
